DSA_FINAL_PROJECT1/pyproject.toml
[project]
name = "dsa-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-agents>=0.6.1",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.66",
]
